% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_interface.R
\name{add_sample_size_metrics_to_df}
\alias{add_sample_size_metrics_to_df}
\title{Helper to add sample size metrics to summary df}
\usage{
add_sample_size_metrics_to_df(df, results)
}
\description{
Helper to add sample size metrics to summary df
}
\keyword{internal}
