% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{compare_sample_sizes}
\alias{compare_sample_sizes}
\title{Compare sample sizes under H0 and H1}
\usage{
compare_sample_sizes(results, sample_size_fn, ...)
}
\arguments{
\item{results}{Output from find_lambda_for_power}

\item{sample_size_fn}{Sample size function}

\item{...}{Additional arguments passed to sample_size_fn}
}
\description{
Compare sample sizes under H0 and H1
}
