% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{create_design_plots}
\alias{create_design_plots}
\title{Create design variable plots}
\usage{
create_design_plots(plot_data, design_vars, boundaries, palette)
}
\description{
Create design variable plots
}
\keyword{internal}
