% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_templates.R
\name{crossover_model_builder}
\alias{crossover_model_builder}
\title{Model builder for crossover CRT}
\usage{
crossover_model_builder(design_params, fixed_params)
}
\description{
Model builder for crossover CRT
}
\keyword{internal}
