% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_design_guide.R
\name{custom_design_help}
\alias{custom_design_help}
\title{Print help on custom design creation}
\usage{
custom_design_help(topic = NULL)
}
\arguments{
\item{topic}{Optional topic: "model_builder", "resources", "cost_fn",
"sample_size_fn", "example", or NULL for overview}
}
\description{
Print help on custom design creation
}
