% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_spec.R
\name{eval_resource}
\alias{eval_resource}
\title{Evaluate a resource formula in an environment}
\usage{
eval_resource(formula, env)
}
\arguments{
\item{formula}{A formula like ~ n_arms * k1 * m1}

\item{env}{A list or environment with variable values}
}
\value{
Numeric result
}
\description{
Evaluate a resource formula in an environment
}
