% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{find_contiguous_blocks}
\alias{find_contiguous_blocks}
\title{Find contiguous blocks of outcomes with extended boundaries}
\usage{
find_contiguous_blocks(z1, outcome)
}
\description{
Find contiguous blocks of outcomes with extended boundaries
}
\keyword{internal}
