% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{find_decision_boundaries}
\alias{find_decision_boundaries}
\title{Find decision boundaries from plot data}
\usage{
find_decision_boundaries(plot_data)
}
\description{
Find decision boundaries from plot data
}
\keyword{internal}
