% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{fixed_crossover_model_builder}
\alias{fixed_crossover_model_builder}
\title{Fixed crossover model builder}
\usage{
fixed_crossover_model_builder(design_params, fixed_params)
}
\description{
Fixed crossover model builder
}
\keyword{internal}
