% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_templates.R
\name{fixed_parallel_model_builder}
\alias{fixed_parallel_model_builder}
\title{Fixed parallel model builder (for comparison)}
\usage{
fixed_parallel_model_builder(design_params, fixed_params)
}
\description{
Fixed parallel model builder (for comparison)
}
\keyword{internal}
