% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{fixed_stepped_wedge_model_builder}
\alias{fixed_stepped_wedge_model_builder}
\title{Fixed stepped-wedge model builder}
\usage{
fixed_stepped_wedge_model_builder(design_params, fixed_params)
}
\description{
Fixed stepped-wedge model builder
}
\keyword{internal}
