% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_spec.R
\name{generate_fixed_cost_fn}
\alias{generate_fixed_cost_fn}
\title{Generate fixed design cost function}
\usage{
generate_fixed_cost_fn(spec)
}
\arguments{
\item{spec}{A crt_design_spec object}
}
\value{
A cost function for single-stage designs
}
\description{
Generate fixed design cost function
}
