% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_spec.R
\name{generate_fixed_model_builder}
\alias{generate_fixed_model_builder}
\title{Generate fixed design model from adaptive spec}
\usage{
generate_fixed_model_builder(spec)
}
\arguments{
\item{spec}{A crt_design_spec object}
}
\value{
A model builder function for single-stage designs
}
\description{
Generate fixed design model from adaptive spec
}
