% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_spec.R
\name{generate_sample_size_fn}
\alias{generate_sample_size_fn}
\title{Generate sample size function from specification}
\usage{
generate_sample_size_fn(spec)
}
\arguments{
\item{spec}{A crt_design_spec object}
}
\value{
A function with signature function(opt_designs, results, rho)
}
\description{
Generate sample size function from specification
}
