% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_interface.R
\name{get_decision_rules}
\alias{get_decision_rules}
\title{Get decision rules from results}
\usage{
get_decision_rules(results, design_index = 1, z1_values = NULL)
}
\arguments{
\item{results}{An adaptive_results object}

\item{design_index}{For exploration results, which design to use}

\item{z1_values}{Optional specific z1 values to evaluate}
}
\value{
Data frame of decision rules
}
\description{
Get decision rules from results
}
