% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_interface.R,
%   R/adaptive_plotting_comparison.R
\name{get_fixed_design_defaults}
\alias{get_fixed_design_defaults}
\title{Get default fixed design configuration for a design type}
\usage{
get_fixed_design_defaults(design, target_power)

get_fixed_design_defaults(design, target_power)
}
\description{
Get default fixed design configuration for a design type

Get default fixed design configuration for a design type
}
\keyword{internal}
