% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_interface.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Get sample size distribution from results}
\usage{
get_sample_size(results, design_index = 1, hypothesis = "H1")
}
\arguments{
\item{results}{An adaptive_results object}

\item{design_index}{For exploration results, which design to use (default: 1)}

\item{hypothesis}{"H1" or "H0"}
}
\value{
A sample size summary object
}
\description{
Get sample size distribution from results
}
