% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_design_space_v2.R
\name{make_stage1_grid}
\alias{make_stage1_grid}
\title{Create a standard (k1, m1) stage 1 grid}
\usage{
make_stage1_grid(k1_values, m1_values)
}
\arguments{
\item{k1_values}{Vector of k1 values (clusters per arm)}

\item{m1_values}{Vector of m1 values (individuals per cluster-period)}
}
\value{
Data frame suitable for explore_design_space
}
\description{
Create a standard (k1, m1) stage 1 grid
}
