% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_templates.R
\name{parallel_model_builder}
\alias{parallel_model_builder}
\title{Model builder for parallel CRT (internal)}
\usage{
parallel_model_builder(design_params, fixed_params)
}
\description{
Model builder for parallel CRT (internal)
}
\keyword{internal}
