% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_interface.R
\name{plot_comparison_multi}
\alias{plot_comparison_multi}
\title{Plot comparison for multiple adaptive designs}
\usage{
plot_comparison_multi(x, type = c("cost_ratio", "savings", "frontier", "all"))
}
\description{
Plot comparison for multiple adaptive designs
}
\keyword{internal}
