% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{plot_design_heatmap}
\alias{plot_design_heatmap}
\title{Create heatmap of design space}
\usage{
plot_design_heatmap(exploration, z_var = "E_N", x_var = NULL, y_var = NULL)
}
\arguments{
\item{exploration}{Output from explore_design_space}

\item{z_var}{Variable for fill colour}

\item{x_var, }{y_var Variables for axes (default: first two stage1 vars)}
}
\description{
Create heatmap of design space
}
