% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{plot_design_space}
\alias{plot_design_space}
\title{Plot design space exploration results}
\usage{
plot_design_space(
  exploration,
  type = c("EN", "Nmax", "clusters", "power", "probabilities", "all"),
  x_var = "power_stage1",
  group_var = NULL
)
}
\arguments{
\item{exploration}{Output from explore_design_space}

\item{type}{"EN", "Nmax", "clusters", "power", "probabilities", or "all"}

\item{x_var}{Variable for x-axis (default: "power_stage1")}

\item{group_var}{Variable for grouping/colour (default: first column of stage1_grid)}
}
\description{
Plot design space exploration results
}
