% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_design_space_v2.R
\name{plot_exploration}
\alias{plot_exploration}
\title{Plot design space exploration results}
\usage{
plot_exploration(
  exploration,
  x_var = NULL,
  y_var = "E_N",
  colour_var = NULL,
  label_var = NULL,
  facet_var = NULL
)
}
\arguments{
\item{exploration}{Output from explore_design_space}

\item{x_var}{Variable for x-axis (default: first column of stage1_grid)}

\item{y_var}{Variable for y-axis (default: "E_N")}

\item{colour_var}{Variable for colour grouping (default: second column of stage1_grid)}

\item{label_var}{Variable for point labels (optional)}

\item{facet_var}{Variable for faceting (optional)}
}
\description{
Plot design space exploration results
}
