% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_design_space_v2.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Plot Pareto frontier}
\usage{
plot_pareto(pareto, all_designs = NULL, objectives = NULL, ...)
}
\arguments{
\item{pareto}{Data frame of Pareto-optimal designs}

\item{all_designs}{Optional data frame of all designs to show dominated points}

\item{objectives}{Named list of objectives used (for axis labels)}

\item{...}{Additional arguments}
}
\description{
Plot Pareto frontier
}
