% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{plot_sample_size_comparison}
\alias{plot_sample_size_comparison}
\title{Plot sample size comparison (H0 vs H1)}
\usage{
plot_sample_size_comparison(ss_compare, type, show_outcomes)
}
\description{
Plot sample size comparison (H0 vs H1)
}
