% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_cluster_trial_v2.R
\name{precompute_models}
\alias{precompute_models}
\title{Precompute models for all designs in grid}
\usage{
precompute_models(
  design_grid,
  model_fn,
  fixed_params,
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{design_grid}{Data frame of design parameters to search over}

\item{model_fn}{Model function with signature function(design_row, fixed_params)}

\item{fixed_params}{List of fixed parameters passed to model_fn}

\item{parallel}{Use parallel computation}

\item{n_cores}{Number of cores (default: all - 1)}
}
\value{
List of model outputs
}
\description{
Precompute models for all designs in grid
}
