% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_spec.R
\name{resolve_weight}
\alias{resolve_weight}
\title{Get cost weight, resolving references to variables}
\usage{
resolve_weight(weight_spec, env)
}
\arguments{
\item{weight_spec}{Weight specification (number or variable name as string)}

\item{env}{Environment with variables}
}
\value{
Numeric weight
}
\description{
Get cost weight, resolving references to variables
}
