% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_design_guide.R
\name{staggered_enrollment_example}
\alias{staggered_enrollment_example}
\title{Create a custom CRT design (copy and modify this template)}
\usage{
staggered_enrollment_example()
}
\description{
Create a custom CRT design (copy and modify this template)
}
