% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_plotting_comparison.R
\name{summarise_pareto}
\alias{summarise_pareto}
\title{Summarise Pareto frontier}
\usage{
summarise_pareto(pareto_df)
}
\arguments{
\item{pareto_df}{Output from find_constrained_pareto}
}
\value{
Summary data frame (invisibly)
}
\description{
Summarise Pareto frontier
}
